reorganizar el proyecto, los filters y Fields ahora estan en la clase de repo correspondiente
