--- src/main/java/com/nutrehogar/sistemacontable/domain/components/TipoCuenta.java
@@ -1,44 +1,0 @@
-package com.nutrehogar.sistemacontable.domain.components;
-
-import lombok.Getter;
-
-/**
- * Enum que representa los tipos de cuentas contables.
- *
- * <p>Cada tipo de cuenta define su comportamiento financiero y su rol en la estructura contable de la empresa.</p>
- *
- * <p><strong>Tipos de Cuenta:</strong></p>
- * <ul>
- *     <li>{@code ACTIVO}: Representa los recursos económicos propiedad de la empresa.</li>
- *     <li>{@code PASIVO}: Representa las obligaciones y deudas de la empresa.</li>
- *     <li>{@code CAPITAL}: Representa los recursos aportados por los propietarios o accionistas.</li>
- *     <li>{@code INGRESO}: Representa los ingresos generados por la empresa.</li>
- *     <li>{@code GASTO}: Representa los gastos incurridos por la empresa.</li>
- *     <li>{@code RESULTADOS}: Representa los resultados financieros netos (utilidades o pérdidas).</li>
- *     <li>{@code PATRIMONIO}: Representa el patrimonio de la empresa, combinando capital, reservas y resultados acumulados.</li>
- * </ul>
- */
-//@Getter
-//public enum TipoCuenta {
-//    ACTIVO("Activo"),
-//    PASIVO("Pasivo"),
-//    CAPITAL("Capital"),
-//    INGRESO("Ingreso"),
-//    GASTO("Gasto"),
-//    RESULTADOS("Resultados"),
-//    PATRIMONIO("Patrimonio");
-//
-//    /**
-//     * Descripción del tipo de cuenta.
-//     */
-//    private final String descripcion;
-//
-//    /**
-//     * Constructor de {@code TipoCuenta}.
-//     *
-//     * @param descripcion Descripción detallada del tipo de cuenta.
-//     */
-//    TipoCuenta(String descripcion) {
-//        this.descripcion = descripcion;
-//    }
-//}
