--- src/main/java/com/nutrehogar/sistemacontable/domain/components/package-info.java
@@ -1,22 +1,0 @@
-/**
- * Paquete que contiene las enumeraciones del dominio de {@code SistemaContable}.
- * <p>
- * Las enumeraciones definen tipos y categorías utilizadas en las entidades, proporcionando un conjunto fijo de valores para propiedades específicas.
- *
- * <p>
- * <strong>Responsabilidades:</strong>
- * <ul>
- *     <li>Definir tipos de documentos contables.</li>
- *     <li>Clasificar las cuentas contables.</li>
- * </ul>
- * </p>
- *
- * <p>
- * <strong>Ejemplos de Enumeraciones:</strong>
- * <ul>
- *     <li>{@code TipoDocumento}</li>
- *     <li>{@code TipoCuenta}</li>
- * </ul>
- * </p>
- */
-package com.nutrehogar.sistemacontable.domain.components;
