+++ src/main/java/com/nutrehogar/sistemacontable/domain/repository/AsientoRepo.java
@@ -1,7 +1,7 @@
-package com.nutrehogar.sistemacontable.persistence.repository;
+package com.nutrehogar.sistemacontable.domain.repository;
 
-import com.nutrehogar.sistemacontable.domain.model.Registro;
-import com.nutrehogar.sistemacontable.persistence.config.HibernateUtil;
+import com.nutrehogar.sistemacontable.domain.model.Asiento;
+import com.nutrehogar.sistemacontable.domain.HibernateUtil;
 import org.hibernate.Session;
 import org.jetbrains.annotations.NotNull;
 
@@ -9,25 +9,25 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
-public class RegistroRepo {
+public class AsientoRepo {
     private static final Session session = HibernateUtil.getSession();
-    private static RegistroRepo instance;
+    private static AsientoRepo instance;
 
-    private RegistroRepo() {
+    private AsientoRepo() {
     }
 
-    public static RegistroRepo getInstance() {
+    public static AsientoRepo getInstance() {
         if (instance == null) {
-            instance = new RegistroRepo();
+            instance = new AsientoRepo();
         }
         return instance;
     }
 
-    public List<Registro> findAll() {
-        List<Registro> Transaccions;
+    public List<Asiento> findAll() {
+        List<Asiento> Transaccions;
         try {
             session.beginTransaction();
-            Transaccions = session.createQuery("from Registro", Registro.class).list();
+            Transaccions = session.createQuery("from Asiento", Asiento.class).list();
             session.getTransaction().commit();
         } catch (Exception e) {
             session.getTransaction().rollback();
@@ -37,35 +37,35 @@ public class RegistroRepo {
         return Transaccions;
     }
 
-    public Registro findById(Integer id) {
-        return session.find(Registro.class, id);
+    public Asiento findById(Integer id) {
+        return session.find(Asiento.class, id);
     }
 
-    public void save(Registro Registro) {
+    public void save(Asiento asiento) {
         try {
             session.beginTransaction();
-            session.persist(Registro);
+            session.persist(asiento);
             session.getTransaction().commit();
         } catch (Exception e) {
             session.getTransaction().rollback();
         }
     }
 
-    public void save(@NotNull List<Registro> registros) {
+    public void save(@NotNull List<Asiento> asientos) {
         try {
             session.beginTransaction();
-            registros.forEach(session::persist);
+            asientos.forEach(session::persist);
             session.getTransaction().commit();
         } catch (Exception e) {
             session.getTransaction().rollback();
         }
     }
 
-    public void delete(Registro registro) {
+    public void delete(Asiento asiento) {
         try {
             session.beginTransaction();
 
-            session.remove(session.contains(registro) ? registro : session.merge(registro));
+            session.remove(session.contains(asiento) ? asiento : session.merge(asiento));
 
             session.getTransaction().commit();
         } catch (Exception e) {
@@ -78,7 +78,7 @@ public class RegistroRepo {
         try {
             session.beginTransaction();
 
-            Optional.ofNullable(findById(id)).ifPresent(session::remove);
+            Optional.ofNullable(session.find(Asiento.class, id)).ifPresent(session::remove);
 
             session.getTransaction().commit();
         } catch (Exception e) {
@@ -87,10 +87,10 @@ public class RegistroRepo {
         }
     }
 
-    public void update(Registro registro) {
+    public void update(Asiento asiento) {
         try {
             session.beginTransaction();
-            session.merge(registro);
+            session.merge(asiento);
             session.getTransaction().commit();
         } catch (Exception e) {
             session.getTransaction().rollback();
