+++ src/main/java/com/nutrehogar/sistemacontable/domain/repository/TipoDocumentoRepo.java
@@ -1,7 +1,7 @@
-package com.nutrehogar.sistemacontable.persistence.repository;
+package com.nutrehogar.sistemacontable.domain.repository;
 
-import com.nutrehogar.sistemacontable.domain.model.Asiento;
-import com.nutrehogar.sistemacontable.persistence.config.HibernateUtil;
+import com.nutrehogar.sistemacontable.domain.model.TipoDocumento;
+import com.nutrehogar.sistemacontable.domain.HibernateUtil;
 import org.hibernate.Session;
 import org.jetbrains.annotations.NotNull;
 
@@ -9,25 +9,25 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
-public class AsientoRepo {
+public class TipoDocumentoRepo {
     private static final Session session = HibernateUtil.getSession();
-    private static AsientoRepo instance;
+    private static TipoDocumentoRepo instance;
 
-    private AsientoRepo() {
+    private TipoDocumentoRepo() {
     }
 
-    public static AsientoRepo getInstance() {
+    public static TipoDocumentoRepo getInstance() {
         if (instance == null) {
-            instance = new AsientoRepo();
+            instance = new TipoDocumentoRepo();
         }
         return instance;
     }
 
-    public List<Asiento> findAll() {
-        List<Asiento> Transaccions;
+    public List<TipoDocumento> findAll() {
+        List<TipoDocumento> Transaccions;
         try {
             session.beginTransaction();
-            Transaccions = session.createQuery("from Asiento", Asiento.class).list();
+            Transaccions = session.createQuery("from TipoDocumento", TipoDocumento.class).list();
             session.getTransaction().commit();
         } catch (Exception e) {
             session.getTransaction().rollback();
@@ -37,35 +37,35 @@ public class AsientoRepo {
         return Transaccions;
     }
 
-    public Asiento findById(Integer id) {
-        return session.find(Asiento.class, id);
+    public TipoDocumento findById(Integer id) {
+        return session.find(TipoDocumento.class, id);
     }
 
-    public void save(Asiento asiento) {
+    public void save(TipoDocumento tipoDocumento) {
         try {
             session.beginTransaction();
-            session.persist(asiento);
+            session.persist(tipoDocumento);
             session.getTransaction().commit();
         } catch (Exception e) {
             session.getTransaction().rollback();
         }
     }
 
-    public void save(@NotNull List<Asiento> asientos) {
+    public void save(@NotNull List<TipoDocumento> tipoDocumentos) {
         try {
             session.beginTransaction();
-            asientos.forEach(session::persist);
+            tipoDocumentos.forEach(session::persist);
             session.getTransaction().commit();
         } catch (Exception e) {
             session.getTransaction().rollback();
         }
     }
 
-    public void delete(Asiento asiento) {
+    public void delete(TipoDocumento tipoDocumento) {
         try {
             session.beginTransaction();
 
-            session.remove(session.contains(asiento) ? asiento : session.merge(asiento));
+            session.remove(session.contains(tipoDocumento) ? tipoDocumento : session.merge(tipoDocumento));
 
             session.getTransaction().commit();
         } catch (Exception e) {
@@ -78,8 +78,7 @@ public class AsientoRepo {
         try {
             session.beginTransaction();
 
-            Optional.ofNullable(session.find(Asiento.class, id)).ifPresent(session::remove);
-
+            Optional.ofNullable(findById(id)).ifPresent(session::remove);
             session.getTransaction().commit();
         } catch (Exception e) {
             session.getTransaction().rollback();
@@ -87,10 +86,10 @@ public class AsientoRepo {
         }
     }
 
-    public void update(Asiento asiento) {
+    public void update(TipoDocumento tipoDocumento) {
         try {
             session.beginTransaction();
-            session.merge(asiento);
+            session.merge(tipoDocumento);
             session.getTransaction().commit();
         } catch (Exception e) {
             session.getTransaction().rollback();
