+++ src/main/java/com/nutrehogar/sistemacontable/domain/repository/package-info.java
@@ -21,4 +21,4 @@
  * </ul>
  * </p>
  */
-package com.nutrehogar.sistemacontable.persistence.repository;
+package com.nutrehogar.sistemacontable.domain.repository;
