--- src/main/java/com/nutrehogar/sistemacontable/domain/util/package-info.java
@@ -1,23 +1,0 @@
-/**
- * Paquete que contiene las clases de filtrado del dominio de {@code SistemaContable}.
- * <p>
- * Estas clases definen los diferentes criterios y periodos para filtrar las transacciones contables y generar informes financieros.
- *
- * <p>
- * <strong>Responsabilidades:</strong>
- * <ul>
- *     <li>Definir criterios de filtrado por fechas, cuentas y otros parámetros.</li>
- *     <li>Facilitar la generación de reportes personalizados.</li>
- * </ul>
- * </p>
- *
- * <p>
- * <strong>Ejemplos de Clases:</strong>
- * <ul>
- *     <li>{@code TransactionFilter}</li>
- *     <li>{@code LibroDiarioFilter}</li>
- *     <li>{@code BalanceComprobacionFilter}</li>
- * </ul>
- * </p>
- */
-package com.nutrehogar.sistemacontable.domain.util;
