--- src/main/java/com/nutrehogar/sistemacontable/persistence/config/Revision.java
@@ -1,4 +1,0 @@
-package com.nutrehogar.sistemacontable.persistence.config;
-
-public class Revision {
-}
