--- src/main/java/com/nutrehogar/sistemacontable/persistence/config/package-info.java
@@ -1,23 +1,0 @@
-/**
- * Paquete que contiene las configuraciones de persistencia de {@code SistemaContable}.
- * <p>
- * Incluye configuraciones relacionadas con Hibernate, como la creación de la SessionFactory y la configuración de las conexiones a la base de datos.
- *
- * <p>
- * <strong>Responsabilidades:</strong>
- * <ul>
- *     <li>Configurar Hibernate y otros frameworks de persistencia.</li>
- *     <li>Gestionar las conexiones a la base de datos.</li>
- *     <li>Definir estrategias de mapeo y generación de esquemas.</li>
- * </ul>
- * </p>
- *
- * <p>
- * <strong>Ejemplos de Clases:</strong>
- * <ul>
- *     <li>{@code HibernateUtil}</li>
- *     <li>{@code DatabaseConfig}</li>
- * </ul>
- * </p>
- */
-package com.nutrehogar.sistemacontable.persistence.config;
