--- src/main/java/com/nutrehogar/sistemacontable/persistence/package-info.java
@@ -1,23 +1,0 @@
-/**
- * Paquete que maneja el acceso a datos en {@code SistemaContable}.
- * <p>
- * Este paquete contiene los repositorios que interactúan con la base de datos y las configuraciones necesarias para la persistencia.
- * Utiliza Hibernate como ORM para mapear las entidades a las tablas de la base de datos.
- *
- * <p>
- * <strong>Ámbito de Desarrollo:</strong>
- * <ul>
- *     <li>Configuración de Hibernate y otras tecnologías de persistencia.</li>
- *     <li>Implementación de repositorios para acceder a los datos.</li>
- * </ul>
- * </p>
- *
- * <p>
- * <strong>Ámbito Financiero:</strong>
- * <ul>
- *     <li>Acceso y manipulación de datos contables almacenados en la base de datos.</li>
- *     <li>Asegurar la integridad y consistencia de los datos financieros.</li>
- * </ul>
- * </p>
- */
-package com.nutrehogar.sistemacontable.persistence;
