+++ src/main/java/com/nutrehogar/sistemacontable/ui/view/ListaCuenta.java
@@ -1,140 +1,133 @@
-
 package com.nutrehogar.sistemacontable.ui.view;
 
 import com.nutrehogar.sistemacontable.domain.model.Cuenta;
 import com.nutrehogar.sistemacontable.domain.model.SubTipoCuenta;
-import com.nutrehogar.sistemacontable.persistence.repository.CuentaRepo;
-import com.nutrehogar.sistemacontable.persistence.repository.SubTipoCuentaRepo;
-import java.awt.Color;
-import java.awt.Component;
-import java.awt.Cursor;
-import java.awt.Dimension;
-import java.awt.Font;
-import java.awt.Graphics;
-import java.awt.Graphics2D;
-import java.awt.Insets;
-import java.awt.RenderingHints;
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
-import java.awt.event.MouseMotionAdapter;
-import java.util.ArrayList;
-import java.util.List;
-import javax.swing.JButton;
-import javax.swing.JComponent;
-import javax.swing.JOptionPane;
-import javax.swing.JPanel;
-import javax.swing.JTable;
+import com.nutrehogar.sistemacontable.domain.repository.CuentaRepo;
+import com.nutrehogar.sistemacontable.domain.repository.SubTipoCuentaRepo;
+
+import javax.swing.*;
 import javax.swing.event.ListSelectionEvent;
 import javax.swing.table.DefaultTableCellRenderer;
 import javax.swing.table.DefaultTableModel;
 import javax.swing.table.JTableHeader;
 import javax.swing.table.TableColumnModel;
+import java.awt.*;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
- *
  * @author charl
  */
 public class ListaCuenta extends javax.swing.JPanel {
- List<SubTipoCuenta> listaSubTipo = new ArrayList<>();
+    List<SubTipoCuenta> listaSubTipo = new ArrayList<>();
 
-        SubTipoCuentaRepo subTipoCuentaRepo = SubTipoCuentaRepo.getInstance();
+    SubTipoCuentaRepo subTipoCuentaRepo = SubTipoCuentaRepo.getInstance();
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JButton btnActualizar;
+    private javax.swing.JButton btnCrear;
+    private javax.swing.JButton btnEliminar;
+    private javax.swing.JLabel jLabel1;
+    private javax.swing.JPanel jPanel1;
+    private javax.swing.JScrollPane jScrollPane1;
+    private javax.swing.JTable tblCuenta;
+
     public ListaCuenta() {
         initComponents();
         estilizarBoton(btnCrear, 120, 50);      // Botón Crear con ancho 120 y alto 50
-estilizarBoton(btnActualizar, 120, 50); // Botón Modificar con ancho 120 y alto 50
-estilizarBoton(btnEliminar, 120, 50);  // Botón Eliminar con ancho 120 y alto 50
+        estilizarBoton(btnActualizar, 120, 50); // Botón Modificar con ancho 120 y alto 50
+        estilizarBoton(btnEliminar, 120, 50);  // Botón Eliminar con ancho 120 y alto 50
 
-       customizeTable();
+        customizeTable();
         tblCuenta.getSelectionModel().addListSelectionListener(this::isSelectRow);
         btnActualizar.setEnabled(false);
 
 
         List<Cuenta> ListaCuenta = CuentaRepo.getInstance().findAll();
-        DefaultTableModel model = (DefaultTableModel)tblCuenta.getModel();
+        DefaultTableModel model = (DefaultTableModel) tblCuenta.getModel();
         for (Cuenta cuenta : ListaCuenta) {
             model.addRow(new Object[]{
-            cuenta.getId(),cuenta.getNombre(),cuenta.getSubTipoCuenta().getNombre(),cuenta.getSubTipoCuenta().getTipoCuenta().getNombre()
-        });
+                    cuenta.getId(), cuenta.getNombre(), cuenta.getSubTipoCuenta().getNombre(), cuenta.getSubTipoCuenta().getTipoCuenta().getNombre()
+            });
         }
-        
 
 
     }
+
     private void estilizarBoton(JButton boton, int ancho, int alto) {
-    // Cambiar tamaño del botón
-    boton.setPreferredSize(new Dimension(ancho, alto));
+        // Cambiar tamaño del botón
+        boton.setPreferredSize(new Dimension(ancho, alto));
 
-    // Configuración básica para eliminar estilos predeterminados
-    boton.setContentAreaFilled(false); // Elimina el fondo predeterminado
-    boton.setFocusPainted(false);      // Elimina el borde de enfoque predeterminado
-    boton.setBorderPainted(false);     // Elimina el borde predeterminado
+        // Configuración básica para eliminar estilos predeterminados
+        boton.setContentAreaFilled(false); // Elimina el fondo predeterminado
+        boton.setFocusPainted(false);      // Elimina el borde de enfoque predeterminado
+        boton.setBorderPainted(false);     // Elimina el borde predeterminado
 
-    // Estilo personalizado del botón
-    boton.setUI(new javax.swing.plaf.basic.BasicButtonUI() {
-        @Override
-        public void paint(Graphics g, JComponent c) {
-            Graphics2D g2d = (Graphics2D) g.create();
-            g2d.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
+        // Estilo personalizado del botón
+        boton.setUI(new javax.swing.plaf.basic.BasicButtonUI() {
+            @Override
+            public void paint(Graphics g, JComponent c) {
+                Graphics2D g2d = (Graphics2D) g.create();
+                g2d.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
 
-            // Dibujar el fondo con bordes redondeados
-            g2d.setColor(new Color(102, 204, 255)); // Color de fondo
-            g2d.fillRoundRect(0, 0, c.getWidth(), c.getHeight(), 25, 25); // Bordes redondeados
+                // Dibujar el fondo con bordes redondeados
+                g2d.setColor(new Color(102, 204, 255)); // Color de fondo
+                g2d.fillRoundRect(0, 0, c.getWidth(), c.getHeight(), 25, 25); // Bordes redondeados
 
-            // Dibujar el borde
-            g2d.setColor(new Color(30, 144, 255)); // Color del borde
-            g2d.drawRoundRect(0, 0, c.getWidth() - 1, c.getHeight() - 1, 25, 25);
+                // Dibujar el borde
+                g2d.setColor(new Color(30, 144, 255)); // Color del borde
+                g2d.drawRoundRect(0, 0, c.getWidth() - 1, c.getHeight() - 1, 25, 25);
 
-            g2d.dispose();
+                g2d.dispose();
 
-            // Llamar al renderizador predeterminado para pintar el texto y el ícono
-            super.paint(g, c);
-        }
-    });
-}
-
-// Personalización de la tabla
-private void customizeTable() {
-    // Cambiar el estilo del encabezado
-    JTableHeader header = tblCuenta.getTableHeader();
-    header.setFont(new Font("Arial", Font.BOLD, 16));
-    header.setBackground(new Color(102, 204, 255));
-    header.setForeground(Color.BLACK);
-
-    // Renderizador para aplicar estilos en todas las filas
-    DefaultTableCellRenderer rowRenderer = new DefaultTableCellRenderer() {
-        @Override
-        public Component getTableCellRendererComponent(JTable table, Object value,
-                                                       boolean isSelected, boolean hasFocus,
-                                                       int row, int column) {
-            Component cell = super.getTableCellRendererComponent(table, value, isSelected, hasFocus, row, column);
-
-            // Alternar colores de las filas
-            if (!isSelected) {
-                setBackground(row % 2 == 0 ? new Color(224, 255, 255) : Color.WHITE); // Alterna entre celeste claro y blanco
-                setForeground(Color.BLACK);
-            } else {
-                setBackground(new Color(30, 144, 255)); // Azul al seleccionar
-                setForeground(Color.WHITE);
+                // Llamar al renderizador predeterminado para pintar el texto y el ícono
+                super.paint(g, c);
             }
-            return cell;
-        }
-    };
-
-    // Asignar el renderizador a todas las columnas de la tabla
-    for (int i = 0; i < tblCuenta.getColumnCount(); i++) {
-        tblCuenta.getColumnModel().getColumn(i).setCellRenderer(rowRenderer);
+        });
     }
 
-    // Ajustar tamaño de texto y altura de las filas
-    tblCuenta.setFont(new Font("Arial", Font.PLAIN, 14));
-    tblCuenta.setRowHeight(25);
+    // Personalización de la tabla
+    private void customizeTable() {
+        // Cambiar el estilo del encabezado
+        JTableHeader header = tblCuenta.getTableHeader();
+        header.setFont(new Font("Arial", Font.BOLD, 16));
+        header.setBackground(new Color(102, 204, 255));
+        header.setForeground(Color.BLACK);
 
-    // Ajustar el ancho de la columna ID si es necesario
-    TableColumnModel columnModel = tblCuenta.getColumnModel();
-    columnModel.getColumn(0).setPreferredWidth(80);
-    columnModel.getColumn(0).setMaxWidth(80);
-    columnModel.getColumn(0).setMinWidth(80);
-}
+        // Renderizador para aplicar estilos en todas las filas
+        DefaultTableCellRenderer rowRenderer = new DefaultTableCellRenderer() {
+            @Override
+            public Component getTableCellRendererComponent(JTable table, Object value,
+                                                           boolean isSelected, boolean hasFocus,
+                                                           int row, int column) {
+                Component cell = super.getTableCellRendererComponent(table, value, isSelected, hasFocus, row, column);
+
+                // Alternar colores de las filas
+                if (!isSelected) {
+                    setBackground(row % 2 == 0 ? new Color(224, 255, 255) : Color.WHITE); // Alterna entre celeste claro y blanco
+                    setForeground(Color.BLACK);
+                } else {
+                    setBackground(new Color(30, 144, 255)); // Azul al seleccionar
+                    setForeground(Color.WHITE);
+                }
+                return cell;
+            }
+        };
+
+        // Asignar el renderizador a todas las columnas de la tabla
+        for (int i = 0; i < tblCuenta.getColumnCount(); i++) {
+            tblCuenta.getColumnModel().getColumn(i).setCellRenderer(rowRenderer);
+        }
+
+        // Ajustar tamaño de texto y altura de las filas
+        tblCuenta.setFont(new Font("Arial", Font.PLAIN, 14));
+        tblCuenta.setRowHeight(25);
+
+        // Ajustar el ancho de la columna ID si es necesario
+        TableColumnModel columnModel = tblCuenta.getColumnModel();
+        columnModel.getColumn(0).setPreferredWidth(80);
+        columnModel.getColumn(0).setMaxWidth(80);
+        columnModel.getColumn(0).setMinWidth(80);
+    }
 
     /**
      * This method is called from within the constructor to initialize the form.
@@ -157,19 +150,19 @@ private void customizeTable() {
         jLabel1.setText("Lista de Cuentas");
 
         tblCuenta.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][] {
+                new Object[][]{
 
-            },
-            new String [] {
-                "ID", "NOMBRE", "TIPO CUENTA", "SUBTIPO CUENTA"
-            }
+                },
+                new String[]{
+                        "ID", "NOMBRE", "TIPO CUENTA", "SUBTIPO CUENTA"
+                }
         ) {
-            Class[] types = new Class [] {
-                java.lang.Integer.class, java.lang.String.class, java.lang.String.class, java.lang.String.class
+            Class[] types = new Class[]{
+                    java.lang.Integer.class, java.lang.String.class, java.lang.String.class, java.lang.String.class
             };
 
             public Class getColumnClass(int columnIndex) {
-                return types [columnIndex];
+                return types[columnIndex];
             }
         });
         jScrollPane1.setViewportView(tblCuenta);
@@ -209,59 +202,60 @@ private void customizeTable() {
         javax.swing.GroupLayout jPanel1Layout = new javax.swing.GroupLayout(jPanel1);
         jPanel1.setLayout(jPanel1Layout);
         jPanel1Layout.setHorizontalGroup(
-            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel1Layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                    .addComponent(btnActualizar, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(btnCrear, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(btnEliminar, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addContainerGap(28, Short.MAX_VALUE))
+                jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jPanel1Layout.createSequentialGroup()
+                                .addContainerGap()
+                                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                        .addComponent(btnActualizar, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                        .addComponent(btnCrear, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                        .addComponent(btnEliminar, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                                .addContainerGap(28, Short.MAX_VALUE))
         );
         jPanel1Layout.setVerticalGroup(
-            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
-                .addGap(19, 19, 19)
-                .addComponent(btnCrear)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 59, Short.MAX_VALUE)
-                .addComponent(btnActualizar)
-                .addGap(63, 63, 63)
-                .addComponent(btnEliminar)
-                .addGap(17, 17, 17))
+                jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
+                                .addGap(19, 19, 19)
+                                .addComponent(btnCrear)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 59, Short.MAX_VALUE)
+                                .addComponent(btnActualizar)
+                                .addGap(63, 63, 63)
+                                .addComponent(btnEliminar)
+                                .addGap(17, 17, 17))
         );
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
         this.setLayout(layout);
         layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(layout.createSequentialGroup()
-                        .addGap(23, 23, 23)
-                        .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 716, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(63, 63, 63)
-                        .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(layout.createSequentialGroup()
-                        .addGap(281, 281, 281)
-                        .addComponent(jLabel1)))
-                .addContainerGap(330, Short.MAX_VALUE))
+                layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(layout.createSequentialGroup()
+                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                        .addGroup(layout.createSequentialGroup()
+                                                .addGap(23, 23, 23)
+                                                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 716, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                .addGap(63, 63, 63)
+                                                .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                        .addGroup(layout.createSequentialGroup()
+                                                .addGap(281, 281, 281)
+                                                .addComponent(jLabel1)))
+                                .addContainerGap(330, Short.MAX_VALUE))
         );
         layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(layout.createSequentialGroup()
-                        .addGap(76, 76, 76)
-                        .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(layout.createSequentialGroup()
-                        .addGap(30, 30, 30)
-                        .addComponent(jLabel1)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(layout.createSequentialGroup()
+                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                        .addGroup(layout.createSequentialGroup()
+                                                .addGap(76, 76, 76)
+                                                .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                        .addGroup(layout.createSequentialGroup()
+                                                .addGap(30, 30, 30)
+                                                .addComponent(jLabel1)
+                                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                                                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
     }// </editor-fold>//GEN-END:initComponents
- public void isSelectRow(ListSelectionEvent e) {
+
+    public void isSelectRow(ListSelectionEvent e) {
         if (!e.getValueIsAdjusting()) {
             int selectRow = tblCuenta.getSelectedRow();
             if (selectRow != -1) {
@@ -272,6 +266,7 @@ private void customizeTable() {
         }
 
     }
+
     private void btnCrearActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnCrearActionPerformed
 
         DefaultTableModel tblCuentaModelo = (DefaultTableModel) tblCuenta.getModel();
@@ -311,7 +306,7 @@ private void customizeTable() {
         }
 
         int confirm = JOptionPane.showConfirmDialog(this, "¿Está seguro de que desea eliminar esta cuenta?",
-            "Confirmar eliminación", JOptionPane.YES_NO_OPTION);
+                "Confirmar eliminación", JOptionPane.YES_NO_OPTION);
         if (confirm != JOptionPane.YES_OPTION) {
             return;
         }
@@ -331,15 +326,5 @@ private void customizeTable() {
             e.printStackTrace();
         }
     }//GEN-LAST:event_btnEliminarActionPerformed
-
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton btnActualizar;
-    private javax.swing.JButton btnCrear;
-    private javax.swing.JButton btnEliminar;
-    private javax.swing.JLabel jLabel1;
-    private javax.swing.JPanel jPanel1;
-    private javax.swing.JScrollPane jScrollPane1;
-    private javax.swing.JTable tblCuenta;
     // End of variables declaration//GEN-END:variables
 }
