+++ src/main/java/com/nutrehogar/sistemacontable/ui/view/MainFrameDePrueba.java
@@ -1,10 +1,4 @@
 package com.nutrehogar.sistemacontable.ui.view;
-import javax.swing.*;
-import javax.swing.table.DefaultTableModel;
-import java.awt.*;
-import java.awt.image.BufferedImage;
-import java.io.FileOutputStream;
-import javax.imageio.ImageIO;
 
 import com.formdev.flatlaf.FlatLightLaf;
 import com.itextpdf.text.Document;
@@ -12,7 +6,13 @@ import com.itextpdf.text.Image;
 import com.itextpdf.text.PageSize;
 import com.itextpdf.text.pdf.PdfWriter;
 import com.nutrehogar.sistemacontable.application.service.PDFService;
-import com.nutrehogar.sistemacontable.ui.view.PlantillaC;
+
+import javax.imageio.ImageIO;
+import javax.swing.*;
+import javax.swing.table.DefaultTableModel;
+import java.awt.*;
+import java.awt.image.BufferedImage;
+import java.io.FileOutputStream;
 
 public class MainFrameDePrueba extends JFrame {
     // Declaración de los componentes
@@ -95,141 +95,142 @@ public class MainFrameDePrueba extends JFrame {
 
         // Añadir el panel de tablas al marco
         add(panelTablas, BorderLayout.CENTER);
-        
+
         // Crear los botones Imprimir y Prueba
         btnDocComprobante = new JButton("Comprobante De Pago");
         btnFormulario = new JButton("Formulario");
         btnPrueba = new JButton("Prueba");
-        
+
         // Crear una instancia de Tstd1
         btnDocComprobante.addActionListener(e -> {
-        PlantillaC Tstd1 = new PlantillaC();
-        PDFService.operacionComprobante(table, txt_ncheque.getText(), txt_fecha.getText(), txt_nodoc.getText(), txt_nombre.getText(), txt_concepto.getText(), Tstd1);
+            PlantillaC Tstd1 = new PlantillaC();
+            PDFService.operacionComprobante(table, txt_ncheque.getText(), txt_fecha.getText(), txt_nodoc.getText(), txt_nombre.getText(), txt_concepto.getText(), Tstd1);
         });
-        
+
         btnFormulario.addActionListener(e -> {
-        PlantillaF Tstd2 = new PlantillaF();
-        PDFService.operacionFormulario(table, tablaex, txt_ncheque.getText(), txt_fecha.getText(), txt_nodoc.getText(), txt_nombre.getText(), txt_concepto.getText(), Tstd2);
+            PlantillaF Tstd2 = new PlantillaF();
+            PDFService.operacionFormulario(table, tablaex, txt_ncheque.getText(), txt_fecha.getText(), txt_nodoc.getText(), txt_nombre.getText(), txt_concepto.getText(), Tstd2);
         });
-        
+
         // Acción del botón "Prueba" para guardar en PDF
         btnPrueba.addActionListener(e -> {
             // Llamar a los métodos directamente
             if (table.isEditing()) {
-            table.getCellEditor().stopCellEditing();
+                table.getCellEditor().stopCellEditing();
             }
             table.clearSelection();
             deseleccionarCuadrosTexto();
             guardarComoPDF(); // Esto ejecutará el método de guardado
         });
-        
+
         // Panel inferior para los botones
         JPanel panelInferior = new JPanel();
         panelInferior.add(btnDocComprobante);
         panelInferior.add(btnFormulario);
-        panelInferior.add(btnPrueba); 
+        panelInferior.add(btnPrueba);
 
         // Añadir el panel inferior al JFrame
         add(panelInferior, BorderLayout.SOUTH);
     }
 
-private void deseleccionarCuadrosTexto() {
-    txt_ncheque.setSelectionStart(0);
-    txt_ncheque.setSelectionEnd(0);
-    txt_fecha.setSelectionStart(0);
-    txt_fecha.setSelectionEnd(0);
-    txt_nombre.setSelectionStart(0);
-    txt_nombre.setSelectionEnd(0);
-    txt_concepto.setSelectionStart(0);
-    txt_concepto.setSelectionEnd(0);
-    txt_nodoc.setSelectionStart(0);
-    txt_nodoc.setSelectionEnd(0);
-    table.clearSelection();
-    
-     // Finalizar la edición actual de la celda (si la hay)
-    if (table.isEditing()) {
-        table.getCellEditor().stopCellEditing();
-    }
-    
-    // Asegurarse de que el renderizador no marque la celda
-    table.getSelectionModel().clearSelection();
-    table.getColumnModel().getSelectionModel().clearSelection();
-}
-
-private void guardarComoPDF() {
-    try {
-        // Obtener el tamaño de la interfaz (sin bordes)
-        Dimension dimension = getSize();
-        BufferedImage imagen = new BufferedImage(dimension.width, dimension.height, BufferedImage.TYPE_INT_RGB);
-        Graphics2D g2d = imagen.createGraphics();
-        
-        // Cambiar el color de fondo (si es necesario)
-        g2d.setColor(getBackground());
-        g2d.fillRect(0, 0, dimension.width, dimension.height);
-
-        // Pintar la interfaz en el BufferedImage
-        paint(g2d);
-        g2d.dispose();
-
-        // Guardar la imagen temporalmente en el sistema de archivos
-        ImageIO.write(imagen, "png", new java.io.File("captura.png"));
-
-        // Usar JFileChooser para seleccionar ubicación
-        JFileChooser fileChooser = new JFileChooser();
-        fileChooser.setDialogTitle("Guardar como PDF");
-        fileChooser.setSelectedFile(new java.io.File("interfaz.pdf"));
-        
-        int userSelection = fileChooser.showSaveDialog(this);
-        
-        if (userSelection == JFileChooser.APPROVE_OPTION) {
-            java.io.File fileToSave = fileChooser.getSelectedFile();
-            
-            // Verificar si el archivo ya existe
-            if (fileToSave.exists()) {
-                // Preguntar al usuario si desea sobrescribir
-                int response = JOptionPane.showConfirmDialog(this, 
-                        "El archivo ya existe. ¿Deseas reemplazarlo?", 
-                        "Confirmar Sobrescritura", 
-                        JOptionPane.YES_NO_OPTION, 
-                        JOptionPane.WARNING_MESSAGE);
-                
-                // Si el usuario elige NO, cancelar el guardado
-                if (response != JOptionPane.YES_OPTION) {
-                    return; // Salir del método sin guardar
-                }
-            }
-
-            // Crear el documento PDF con tamaño Carta (Letter)
-            Document documento = new Document(new com.itextpdf.text.Rectangle(0, 0, 612, 792)); // 612x792 puntos para Carta (8.5"x11")
-            PdfWriter.getInstance(documento, new FileOutputStream(fileToSave));
-            documento.open();
-
-            // Añadir la imagen al documento PDF
-            Image imagenPDF = Image.getInstance("captura.png");
-            imagenPDF.scaleToFit(PageSize.LETTER.getWidth(), PageSize.LETTER.getHeight());
-            imagenPDF.setAlignment(Image.ALIGN_CENTER);  // Centrar imagen
-            documento.add(imagenPDF);
-
-            // Cerrar el documento
-            documento.close();
-
-            JOptionPane.showMessageDialog(this, "PDF guardado exitosamente en " + fileToSave.getAbsolutePath());
-        }
-        
-    } catch (Exception ex) {
-        ex.printStackTrace();
-        JOptionPane.showMessageDialog(this, "Error al guardar en PDF: " + ex.getMessage());
-    }
-}
-
     public static void main(String[] args) {
         try {
             UIManager.setLookAndFeel(new FlatLightLaf());
-        } catch(Exception ex) {
+        } catch (Exception ex) {
             System.err.println("Failed to initialize LaF");
-        }        SwingUtilities.invokeLater(() -> {
+        }
+        SwingUtilities.invokeLater(() -> {
             MainFrameDePrueba frame = new MainFrameDePrueba();
             frame.setVisible(true);
         });
+    }
+
+    private void deseleccionarCuadrosTexto() {
+        txt_ncheque.setSelectionStart(0);
+        txt_ncheque.setSelectionEnd(0);
+        txt_fecha.setSelectionStart(0);
+        txt_fecha.setSelectionEnd(0);
+        txt_nombre.setSelectionStart(0);
+        txt_nombre.setSelectionEnd(0);
+        txt_concepto.setSelectionStart(0);
+        txt_concepto.setSelectionEnd(0);
+        txt_nodoc.setSelectionStart(0);
+        txt_nodoc.setSelectionEnd(0);
+        table.clearSelection();
+
+        // Finalizar la edición actual de la celda (si la hay)
+        if (table.isEditing()) {
+            table.getCellEditor().stopCellEditing();
+        }
+
+        // Asegurarse de que el renderizador no marque la celda
+        table.getSelectionModel().clearSelection();
+        table.getColumnModel().getSelectionModel().clearSelection();
+    }
+
+    private void guardarComoPDF() {
+        try {
+            // Obtener el tamaño de la interfaz (sin bordes)
+            Dimension dimension = getSize();
+            BufferedImage imagen = new BufferedImage(dimension.width, dimension.height, BufferedImage.TYPE_INT_RGB);
+            Graphics2D g2d = imagen.createGraphics();
+
+            // Cambiar el color de fondo (si es necesario)
+            g2d.setColor(getBackground());
+            g2d.fillRect(0, 0, dimension.width, dimension.height);
+
+            // Pintar la interfaz en el BufferedImage
+            paint(g2d);
+            g2d.dispose();
+
+            // Guardar la imagen temporalmente en el sistema de archivos
+            ImageIO.write(imagen, "png", new java.io.File("captura.png"));
+
+            // Usar JFileChooser para seleccionar ubicación
+            JFileChooser fileChooser = new JFileChooser();
+            fileChooser.setDialogTitle("Guardar como PDF");
+            fileChooser.setSelectedFile(new java.io.File("interfaz.pdf"));
+
+            int userSelection = fileChooser.showSaveDialog(this);
+
+            if (userSelection == JFileChooser.APPROVE_OPTION) {
+                java.io.File fileToSave = fileChooser.getSelectedFile();
+
+                // Verificar si el archivo ya existe
+                if (fileToSave.exists()) {
+                    // Preguntar al usuario si desea sobrescribir
+                    int response = JOptionPane.showConfirmDialog(this,
+                            "El archivo ya existe. ¿Deseas reemplazarlo?",
+                            "Confirmar Sobrescritura",
+                            JOptionPane.YES_NO_OPTION,
+                            JOptionPane.WARNING_MESSAGE);
+
+                    // Si el usuario elige NO, cancelar el guardado
+                    if (response != JOptionPane.YES_OPTION) {
+                        return; // Salir del método sin guardar
+                    }
+                }
+
+                // Crear el documento PDF con tamaño Carta (Letter)
+                Document documento = new Document(new com.itextpdf.text.Rectangle(0, 0, 612, 792)); // 612x792 puntos para Carta (8.5"x11")
+                PdfWriter.getInstance(documento, new FileOutputStream(fileToSave));
+                documento.open();
+
+                // Añadir la imagen al documento PDF
+                Image imagenPDF = Image.getInstance("captura.png");
+                imagenPDF.scaleToFit(PageSize.LETTER.getWidth(), PageSize.LETTER.getHeight());
+                imagenPDF.setAlignment(Image.ALIGN_CENTER);  // Centrar imagen
+                documento.add(imagenPDF);
+
+                // Cerrar el documento
+                documento.close();
+
+                JOptionPane.showMessageDialog(this, "PDF guardado exitosamente en " + fileToSave.getAbsolutePath());
+            }
+
+        } catch (Exception ex) {
+            ex.printStackTrace();
+            JOptionPane.showMessageDialog(this, "Error al guardar en PDF: " + ex.getMessage());
+        }
     }
 }
