+++ src/main/java/com/nutrehogar/sistemacontable/ui/view/PlantillaF.java
@@ -8,255 +8,312 @@ import com.itextpdf.text.Document;
 import com.itextpdf.text.Image;
 import com.itextpdf.text.PageSize;
 import com.itextpdf.text.pdf.PdfWriter;
-import java.awt.Color;
-import java.awt.Component;
-import java.awt.Dimension;
-import java.awt.Font;
-import java.awt.Graphics2D;
+
+import javax.imageio.ImageIO;
+import javax.swing.*;
+import javax.swing.table.DefaultTableCellRenderer;
+import javax.swing.table.DefaultTableModel;
+import javax.swing.table.JTableHeader;
+import java.awt.*;
 import java.awt.image.BufferedImage;
 import java.io.FileOutputStream;
 import java.text.SimpleDateFormat;
 import java.util.Date;
-import javax.imageio.ImageIO;
-import javax.swing.JFileChooser;
-import javax.swing.JOptionPane;
-import javax.swing.JScrollPane;
-import javax.swing.JTable;
-import javax.swing.SwingConstants;
-import javax.swing.table.DefaultTableCellRenderer;
-import javax.swing.table.DefaultTableModel;
-import javax.swing.table.JTableHeader;
 
 public class PlantillaF extends javax.swing.JFrame {
     private JTable tabla;
     private JTable tablaex;
     private JTable tablacom;
-    public void setLabelsText(String cheque, String fecha, String noDoc, String nombre, String concepto, String sumaHaber) {
-    lab_ncheque.setText(cheque);
-    lab_nfecha.setText(fecha);
-    lab_ndoc.setText(noDoc);
-    lab_nnombre.setText(nombre);
-    lab_nconcepto.setText(concepto);
-    lab_nmonto.setText(sumaHaber);
-}
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JButton jButton1;
+    private javax.swing.JButton jButton2;
 
-    public void exportarPantallaAPDF() {
-    try {
-        // Obtener el tamaño de la interfaz (sin bordes)
-        Dimension dimension = getSize();
-        BufferedImage imagen = new BufferedImage(dimension.width, dimension.height, BufferedImage.TYPE_INT_RGB);
-        Graphics2D g2d = imagen.createGraphics();
-        
-        // Cambiar el color de fondo (si es necesario)
-        g2d.setColor(getBackground());
-        g2d.fillRect(0, 0, dimension.width, dimension.height);
+    ;
+    private javax.swing.JLabel jLabel1;
+    private javax.swing.JLabel jLabel10;
+    private javax.swing.JLabel jLabel11;
+    private javax.swing.JLabel jLabel13;
+    private javax.swing.JLabel jLabel14;
+    private javax.swing.JLabel jLabel15;
+    private javax.swing.JLabel jLabel16;
+    private javax.swing.JLabel jLabel17;
+    private javax.swing.JLabel jLabel18;
+    private javax.swing.JLabel jLabel2;
+    private javax.swing.JLabel jLabel3;
+    private javax.swing.JLabel jLabel4;
+    private javax.swing.JLabel jLabel5;
+    private javax.swing.JLabel jLabel6;
+    private javax.swing.JLabel jLabel7;
+    private javax.swing.JLabel jLabel8;
+    private javax.swing.JScrollPane jScrollPane2;
+    private javax.swing.JTextField lab_ncheque;
+    private javax.swing.JTextArea lab_nconcepto;
+    private javax.swing.JTextField lab_ndoc;
+    private javax.swing.JTextField lab_nfecha;
+    private javax.swing.JTextField lab_nmonto;
+    private javax.swing.JTextField lab_nnombre;
+    public PlantillaF() {
+        initComponents();
+        getContentPane().setBackground(Color.WHITE);
+        setSize(612, 792);
+        setLocationRelativeTo(null); // Centrar en pantalla
+        setLayout(null);
 
-        // Pintar la interfaz en el BufferedImage
-        paint(g2d);
-        g2d.dispose();
+        // Crear el modelo de la tabla original sin filas iniciales
+        DefaultTableModel modelo = new DefaultTableModel(new String[]{"Referencia", "Código", "Debe", "Haber"}, 0);
 
-        // Guardar la imagen temporalmente en el sistema de archivos
-        ImageIO.write(imagen, "png", new java.io.File("captura.png"));
-        setVisible(false);
+        // Crear la tabla original con el modelo
+        tabla = new JTable(modelo);
+        tabla.setRowHeight(16); // Ajustar la altura de las filas
+        tabla.setEnabled(false); // Deshabilitar edición si es necesario
 
-        // Usar JFileChooser para seleccionar ubicación
-        JFileChooser fileChooser = new JFileChooser();
-        fileChooser.setDialogTitle("Guardar como PDF");
-        
-        // Crear un formato de fecha para el nombre del archivo
-        SimpleDateFormat sdf = new SimpleDateFormat("yyyyMMdd"); // Formato de fecha (AñoMesDía)
-        String fechaFormateada = sdf.format(new Date());
+        // Ajustar la altura del encabezado de la tabla
+        JTableHeader header = tabla.getTableHeader();
+        header.setPreferredSize(new Dimension(header.getWidth(), 32));
 
-        //Configurar el nombre del archivo con la palabra "comprobante" y la fecha
-        fileChooser.setSelectedFile(new java.io.File("formulario_" + fechaFormateada + ".pdf"));
-        
-        int userSelection = fileChooser.showSaveDialog(this);
-        
-        if (userSelection == JFileChooser.APPROVE_OPTION) {
-            java.io.File fileToSave = fileChooser.getSelectedFile();
-            
-            // Verificar si el archivo ya existe y mostrar advertencia
-            if (fileToSave.exists()) {
-                int response = JOptionPane.showConfirmDialog(this, 
-                        "El archivo ya existe. ¿Deseas reemplazarlo?", 
-                        "Confirmar Sobrescritura", 
-                        JOptionPane.YES_NO_OPTION, 
-                        JOptionPane.WARNING_MESSAGE);
-                
-                if (response != JOptionPane.YES_OPTION) {
-                    return; // Salir del método sin guardar
+        // Configurar el ancho de las columnas de la tabla original
+        tabla.getColumnModel().getColumn(0).setPreferredWidth(130); // Referencia
+        tabla.getColumnModel().getColumn(1).setPreferredWidth(50);  // Código
+        tabla.getColumnModel().getColumn(2).setPreferredWidth(50);  // Debe
+        tabla.getColumnModel().getColumn(3).setPreferredWidth(50);  // Haber
+
+        // Crear un renderizador personalizado para combinar alineación y formato
+        DefaultTableCellRenderer customRenderer = new DefaultTableCellRenderer() {
+            @Override
+            public Component getTableCellRendererComponent(JTable table, Object value, boolean isSelected, boolean hasFocus, int row, int column) {
+                Component comp = super.getTableCellRendererComponent(table, value, isSelected, hasFocus, row, column);
+
+                // Alineación condicional
+                switch (column) {
+                    case 0:
+                        // Alinear la última celda de la columna "Referencia" a la derecha
+                        setHorizontalAlignment(row == table.getRowCount() - 1 ? SwingConstants.RIGHT : SwingConstants.LEFT);
+                        break;
+                    case 1:
+                        // Alinear "Código" al centro
+                        setHorizontalAlignment(SwingConstants.CENTER);
+                        break;
+                    default:
+                        // Alinear "Debe" y "Haber" a la derecha
+                        setHorizontalAlignment(SwingConstants.RIGHT);
+                        break;
+                }
+
+                // Aplicar negrita a la última fila
+                if (row == table.getRowCount() - 1) {
+                    comp.setFont(comp.getFont().deriveFont(Font.BOLD));
+                } else {
+                    comp.setFont(comp.getFont().deriveFont(Font.PLAIN));
+                }
+
+                return comp;
+            }
+        };
+
+        // Asignar el renderizador a todas las columnas
+        for (int i = 0; i < tabla.getColumnCount(); i++) {
+            tabla.getColumnModel().getColumn(i).setCellRenderer(customRenderer);
+        }
+
+        // Crear un JScrollPane para la tabla original
+        JScrollPane scrollPane = new JScrollPane(tabla);
+        scrollPane.setBounds(304, 215, 269, 338); // Ajustar posición y tamaño
+        scrollPane.setHorizontalScrollBarPolicy(JScrollPane.HORIZONTAL_SCROLLBAR_NEVER);
+        scrollPane.setVerticalScrollBarPolicy(JScrollPane.VERTICAL_SCROLLBAR_NEVER);
+
+        // Crear el modelo para la tabla "tablacom"
+        DefaultTableModel modeloCom = new DefaultTableModel(new String[]{"No. Cheque o Comp."}, 0);
+
+        // Crear la tabla "tablacom" con el modelo
+        tablacom = new JTable(modeloCom);
+        tablacom.setRowHeight(16); // Ajustar la altura de las filas
+        tablacom.setEnabled(false); // Deshabilitar edición si es necesario
+        tablacom.getColumnModel().getColumn(0).setPreferredWidth(80); // Ajustar ancho de la columna
+
+        // Crear un renderizador personalizado para la segunda tabla
+        DefaultTableCellRenderer customRenderer2 = new DefaultTableCellRenderer() {
+            @Override
+            public Component getTableCellRendererComponent(JTable table, Object value, boolean isSelected, boolean hasFocus, int row, int column) {
+                Component comp = super.getTableCellRendererComponent(table, value, isSelected, hasFocus, row, column);
+                setHorizontalAlignment(row == table.getRowCount() - 1 ? SwingConstants.RIGHT : SwingConstants.LEFT);
+
+                // Aplicar negrita a la última fila
+                if (row == table.getRowCount() - 1) {
+                    comp.setFont(comp.getFont().deriveFont(Font.BOLD));
+                } else {
+                    comp.setFont(comp.getFont().deriveFont(Font.PLAIN));
+                }
+
+                return comp;
+            }
+        };
+
+        // Asignar el renderizador a la única columna de la segunda tabla
+        tablacom.getColumnModel().getColumn(0).setCellRenderer(customRenderer2);
+
+        // Ajustar la altura del encabezado de la tabla "tablacom"
+        JTableHeader header2 = tablacom.getTableHeader();
+        header2.setPreferredSize(new Dimension(header2.getWidth(), 32));
+
+        // Crear un JScrollPane para la tabla "tablacom"
+        JScrollPane scrollPaneCom = new JScrollPane(tablacom);
+        scrollPaneCom.setBounds(224, 215, 80, 338); // Ajustar posición y tamaño
+        scrollPaneCom.setHorizontalScrollBarPolicy(JScrollPane.HORIZONTAL_SCROLLBAR_NEVER);
+        scrollPaneCom.setVerticalScrollBarPolicy(JScrollPane.VERTICAL_SCROLLBAR_NEVER);
+
+        // Crear el modelo para la nueva tabla "tablaex"
+        DefaultTableModel modeloEx = new DefaultTableModel(new String[]{"Fecha", "No. Doc", "Tipo de Doc."}, 0);
+
+        // Crear la nueva tabla "tablaex" con el modelo
+        tablaex = new JTable(modeloEx);
+        tablaex.setRowHeight(16); // Ajustar la altura de las filas
+        tablaex.setEnabled(false); // Deshabilitar edición si es necesario
+        tablaex.getColumnModel().getColumn(0).setPreferredWidth(65); // Ajustar ancho de la columna
+        tablaex.getColumnModel().getColumn(2).setPreferredWidth(85); // Ajustar ancho de la columna
+
+        // Ajustar la altura del encabezado de la tabla "tablaex"
+        JTableHeader header3 = tablaex.getTableHeader();
+        header3.setPreferredSize(new Dimension(header3.getWidth(), 32));
+
+        // Crear un JScrollPane para la tabla "tablaex"
+        JScrollPane scrollPaneEx = new JScrollPane(tablaex);
+        scrollPaneEx.setBounds(24, 215, 200, 322); // Ajustar posición y tamaño
+        scrollPaneEx.setHorizontalScrollBarPolicy(JScrollPane.HORIZONTAL_SCROLLBAR_NEVER);
+        scrollPaneEx.setVerticalScrollBarPolicy(JScrollPane.VERTICAL_SCROLLBAR_NEVER);
+
+        // Agregar los JScrollPane al JFrame
+        add(scrollPaneEx); // Tabla a la izquierda
+        add(scrollPaneCom); // Tabla al centro
+        add(scrollPane);    // Tabla a la derecha
+    }
+
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String args[]) {
+        /* Set the Nimbus look and feel */
+        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
+        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
+         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html
+         */
+        try {
+            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
+                if ("Nimbus".equals(info.getName())) {
+                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
+                    break;
                 }
             }
-            
-            // Crear el documento PDF con tamaño Letter (8.5"x11")
-            Document documento = new Document(PageSize.LETTER);
-            PdfWriter.getInstance(documento, new FileOutputStream(fileToSave));
-            documento.open();
-
-            // Añadir la imagen al documento PDF y posicionarla en la parte superior
-            Image imagenPDF = Image.getInstance("captura.png");
-            imagenPDF.scaleToFit(PageSize.LETTER.getWidth(), PageSize.LETTER.getHeight());
-            imagenPDF.setAbsolutePosition(0, PageSize.LETTER.getHeight() - imagenPDF.getScaledHeight()); // Posicionar en la parte superior
-            documento.add(imagenPDF);
-
-            // Cerrar el documento
-            documento.close();
-
-            JOptionPane.showMessageDialog(this, "PDF guardado exitosamente en " + fileToSave.getAbsolutePath());
+        } catch (ClassNotFoundException ex) {
+            java.util.logging.Logger.getLogger(PlantillaF.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (InstantiationException ex) {
+            java.util.logging.Logger.getLogger(PlantillaF.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (IllegalAccessException ex) {
+            java.util.logging.Logger.getLogger(PlantillaF.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
+            java.util.logging.Logger.getLogger(PlantillaF.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         }
-    } catch (Exception ex) {
-        ex.printStackTrace();
-        JOptionPane.showMessageDialog(this, "Error al guardar el PDF: " + ex.getMessage(), "Error", JOptionPane.ERROR_MESSAGE);
-    }
-};
-    
-  public DefaultTableModel getTablaexModel() {
-    return (DefaultTableModel) tablaex.getModel();
-}
-    
-  public DefaultTableModel getTablacomModel() {
-    return (DefaultTableModel) tablacom.getModel();
-}
-    
-  public PlantillaF() {
-    initComponents();
-    getContentPane().setBackground(Color.WHITE);
-    setSize(612, 792);
-    setLocationRelativeTo(null); // Centrar en pantalla
-    setLayout(null);
+        //</editor-fold>
+        //</editor-fold>
 
-    // Crear el modelo de la tabla original sin filas iniciales
-    DefaultTableModel modelo = new DefaultTableModel(new String[]{"Referencia", "Código", "Debe", "Haber"}, 0);
-
-    // Crear la tabla original con el modelo
-    tabla = new JTable(modelo);
-    tabla.setRowHeight(16); // Ajustar la altura de las filas
-    tabla.setEnabled(false); // Deshabilitar edición si es necesario
-
-    // Ajustar la altura del encabezado de la tabla
-    JTableHeader header = tabla.getTableHeader();
-    header.setPreferredSize(new Dimension(header.getWidth(), 32));
-
-    // Configurar el ancho de las columnas de la tabla original
-    tabla.getColumnModel().getColumn(0).setPreferredWidth(130); // Referencia
-    tabla.getColumnModel().getColumn(1).setPreferredWidth(50);  // Código
-    tabla.getColumnModel().getColumn(2).setPreferredWidth(50);  // Debe
-    tabla.getColumnModel().getColumn(3).setPreferredWidth(50);  // Haber
-
-    // Crear un renderizador personalizado para combinar alineación y formato
-    DefaultTableCellRenderer customRenderer = new DefaultTableCellRenderer() {
-        @Override
-        public Component getTableCellRendererComponent(JTable table, Object value, boolean isSelected, boolean hasFocus, int row, int column) {
-            Component comp = super.getTableCellRendererComponent(table, value, isSelected, hasFocus, row, column);
-
-            // Alineación condicional
-            switch (column) {
-                case 0:
-                    // Alinear la última celda de la columna "Referencia" a la derecha
-                    setHorizontalAlignment(row == table.getRowCount() - 1 ? SwingConstants.RIGHT : SwingConstants.LEFT);
-                    break;
-                case 1:
-                    // Alinear "Código" al centro
-                    setHorizontalAlignment(SwingConstants.CENTER);
-                    break;
-                default:
-                    // Alinear "Debe" y "Haber" a la derecha
-                    setHorizontalAlignment(SwingConstants.RIGHT);
-                    break;
+        /* Create and display the form */
+        java.awt.EventQueue.invokeLater(new Runnable() {
+            public void run() {
+                new PlantillaF().setVisible(true);
             }
-
-            // Aplicar negrita a la última fila
-            if (row == table.getRowCount() - 1) {
-                comp.setFont(comp.getFont().deriveFont(Font.BOLD));
-            } else {
-                comp.setFont(comp.getFont().deriveFont(Font.PLAIN));
-            }
-
-            return comp;
-        }
-    };
-    
-    // Asignar el renderizador a todas las columnas
-    for (int i = 0; i < tabla.getColumnCount(); i++) {
-        tabla.getColumnModel().getColumn(i).setCellRenderer(customRenderer);
+        });
     }
-    
-    // Crear un JScrollPane para la tabla original
-    JScrollPane scrollPane = new JScrollPane(tabla);
-    scrollPane.setBounds(304, 215, 269, 338); // Ajustar posición y tamaño
-    scrollPane.setHorizontalScrollBarPolicy(JScrollPane.HORIZONTAL_SCROLLBAR_NEVER);
-    scrollPane.setVerticalScrollBarPolicy(JScrollPane.VERTICAL_SCROLLBAR_NEVER);
 
-    // Crear el modelo para la tabla "tablacom"
-    DefaultTableModel modeloCom = new DefaultTableModel(new String[]{"No. Cheque o Comp."}, 0);
+    public void setLabelsText(String cheque, String fecha, String noDoc, String nombre, String concepto, String sumaHaber) {
+        lab_ncheque.setText(cheque);
+        lab_nfecha.setText(fecha);
+        lab_ndoc.setText(noDoc);
+        lab_nnombre.setText(nombre);
+        lab_nconcepto.setText(concepto);
+        lab_nmonto.setText(sumaHaber);
+    }
 
-    // Crear la tabla "tablacom" con el modelo
-    tablacom = new JTable(modeloCom);
-    tablacom.setRowHeight(16); // Ajustar la altura de las filas
-    tablacom.setEnabled(false); // Deshabilitar edición si es necesario
-    tablacom.getColumnModel().getColumn(0).setPreferredWidth(80); // Ajustar ancho de la columna
+    public void exportarPantallaAPDF() {
+        try {
+            // Obtener el tamaño de la interfaz (sin bordes)
+            Dimension dimension = getSize();
+            BufferedImage imagen = new BufferedImage(dimension.width, dimension.height, BufferedImage.TYPE_INT_RGB);
+            Graphics2D g2d = imagen.createGraphics();
 
-    // Crear un renderizador personalizado para la segunda tabla
-    DefaultTableCellRenderer customRenderer2 = new DefaultTableCellRenderer() {
-    @Override
-    public Component getTableCellRendererComponent(JTable table, Object value, boolean isSelected, boolean hasFocus, int row, int column) {
-        Component comp = super.getTableCellRendererComponent(table, value, isSelected, hasFocus, row, column);
-        setHorizontalAlignment(row == table.getRowCount() - 1 ? SwingConstants.RIGHT : SwingConstants.LEFT);
+            // Cambiar el color de fondo (si es necesario)
+            g2d.setColor(getBackground());
+            g2d.fillRect(0, 0, dimension.width, dimension.height);
 
-        // Aplicar negrita a la última fila
-        if (row == table.getRowCount() - 1) {
-            comp.setFont(comp.getFont().deriveFont(Font.BOLD));
-        } else {
-            comp.setFont(comp.getFont().deriveFont(Font.PLAIN));
+            // Pintar la interfaz en el BufferedImage
+            paint(g2d);
+            g2d.dispose();
+
+            // Guardar la imagen temporalmente en el sistema de archivos
+            ImageIO.write(imagen, "png", new java.io.File("captura.png"));
+            setVisible(false);
+
+            // Usar JFileChooser para seleccionar ubicación
+            JFileChooser fileChooser = new JFileChooser();
+            fileChooser.setDialogTitle("Guardar como PDF");
+
+            // Crear un formato de fecha para el nombre del archivo
+            SimpleDateFormat sdf = new SimpleDateFormat("yyyyMMdd"); // Formato de fecha (AñoMesDía)
+            String fechaFormateada = sdf.format(new Date());
+
+            //Configurar el nombre del archivo con la palabra "comprobante" y la fecha
+            fileChooser.setSelectedFile(new java.io.File("formulario_" + fechaFormateada + ".pdf"));
+
+            int userSelection = fileChooser.showSaveDialog(this);
+
+            if (userSelection == JFileChooser.APPROVE_OPTION) {
+                java.io.File fileToSave = fileChooser.getSelectedFile();
+
+                // Verificar si el archivo ya existe y mostrar advertencia
+                if (fileToSave.exists()) {
+                    int response = JOptionPane.showConfirmDialog(this,
+                            "El archivo ya existe. ¿Deseas reemplazarlo?",
+                            "Confirmar Sobrescritura",
+                            JOptionPane.YES_NO_OPTION,
+                            JOptionPane.WARNING_MESSAGE);
+
+                    if (response != JOptionPane.YES_OPTION) {
+                        return; // Salir del método sin guardar
+                    }
+                }
+
+                // Crear el documento PDF con tamaño Letter (8.5"x11")
+                Document documento = new Document(PageSize.LETTER);
+                PdfWriter.getInstance(documento, new FileOutputStream(fileToSave));
+                documento.open();
+
+                // Añadir la imagen al documento PDF y posicionarla en la parte superior
+                Image imagenPDF = Image.getInstance("captura.png");
+                imagenPDF.scaleToFit(PageSize.LETTER.getWidth(), PageSize.LETTER.getHeight());
+                imagenPDF.setAbsolutePosition(0, PageSize.LETTER.getHeight() - imagenPDF.getScaledHeight()); // Posicionar en la parte superior
+                documento.add(imagenPDF);
+
+                // Cerrar el documento
+                documento.close();
+
+                JOptionPane.showMessageDialog(this, "PDF guardado exitosamente en " + fileToSave.getAbsolutePath());
+            }
+        } catch (Exception ex) {
+            ex.printStackTrace();
+            JOptionPane.showMessageDialog(this, "Error al guardar el PDF: " + ex.getMessage(), "Error", JOptionPane.ERROR_MESSAGE);
         }
+    }
 
-        return comp;
-        }
-    };
+    public DefaultTableModel getTablaexModel() {
+        return (DefaultTableModel) tablaex.getModel();
+    }
 
-    // Asignar el renderizador a la única columna de la segunda tabla
-    tablacom.getColumnModel().getColumn(0).setCellRenderer(customRenderer2);
-    
-    // Ajustar la altura del encabezado de la tabla "tablacom"
-    JTableHeader header2 = tablacom.getTableHeader();
-    header2.setPreferredSize(new Dimension(header2.getWidth(), 32));
+    public DefaultTableModel getTablacomModel() {
+        return (DefaultTableModel) tablacom.getModel();
+    }
 
-    // Crear un JScrollPane para la tabla "tablacom"
-    JScrollPane scrollPaneCom = new JScrollPane(tablacom);
-    scrollPaneCom.setBounds(224, 215, 80, 338); // Ajustar posición y tamaño
-    scrollPaneCom.setHorizontalScrollBarPolicy(JScrollPane.HORIZONTAL_SCROLLBAR_NEVER);
-    scrollPaneCom.setVerticalScrollBarPolicy(JScrollPane.VERTICAL_SCROLLBAR_NEVER);
-
-    // Crear el modelo para la nueva tabla "tablaex"
-    DefaultTableModel modeloEx = new DefaultTableModel(new String[]{"Fecha", "No. Doc", "Tipo de Doc."}, 0);
-
-    // Crear la nueva tabla "tablaex" con el modelo
-    tablaex = new JTable(modeloEx);
-    tablaex.setRowHeight(16); // Ajustar la altura de las filas
-    tablaex.setEnabled(false); // Deshabilitar edición si es necesario
-    tablaex.getColumnModel().getColumn(0).setPreferredWidth(65); // Ajustar ancho de la columna
-    tablaex.getColumnModel().getColumn(2).setPreferredWidth(85); // Ajustar ancho de la columna
-
-    // Ajustar la altura del encabezado de la tabla "tablaex"
-    JTableHeader header3 = tablaex.getTableHeader();
-    header3.setPreferredSize(new Dimension(header3.getWidth(), 32));
-
-    // Crear un JScrollPane para la tabla "tablaex"
-    JScrollPane scrollPaneEx = new JScrollPane(tablaex);
-    scrollPaneEx.setBounds(24, 215, 200, 322); // Ajustar posición y tamaño
-    scrollPaneEx.setHorizontalScrollBarPolicy(JScrollPane.HORIZONTAL_SCROLLBAR_NEVER);
-    scrollPaneEx.setVerticalScrollBarPolicy(JScrollPane.VERTICAL_SCROLLBAR_NEVER);
-    
-    // Agregar los JScrollPane al JFrame
-    add(scrollPaneEx); // Tabla a la izquierda
-    add(scrollPaneCom); // Tabla al centro
-    add(scrollPane);    // Tabla a la derecha
-}  
-    
     public DefaultTableModel getTableModel() {
-    return (DefaultTableModel) tabla.getModel(); 
+        return (DefaultTableModel) tabla.getModel();
     }
-    
+
     @SuppressWarnings("unchecked")
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
@@ -479,68 +536,5 @@ public class PlantillaF extends javax.swing.JFrame {
     private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
         // TODO add your handling code here:
     }//GEN-LAST:event_jButton1ActionPerformed
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(PlantillaF.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(PlantillaF.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(PlantillaF.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(PlantillaF.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new PlantillaF().setVisible(true);
-            }
-        });
-    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton jButton1;
-    private javax.swing.JButton jButton2;
-    private javax.swing.JLabel jLabel1;
-    private javax.swing.JLabel jLabel10;
-    private javax.swing.JLabel jLabel11;
-    private javax.swing.JLabel jLabel13;
-    private javax.swing.JLabel jLabel14;
-    private javax.swing.JLabel jLabel15;
-    private javax.swing.JLabel jLabel16;
-    private javax.swing.JLabel jLabel17;
-    private javax.swing.JLabel jLabel18;
-    private javax.swing.JLabel jLabel2;
-    private javax.swing.JLabel jLabel3;
-    private javax.swing.JLabel jLabel4;
-    private javax.swing.JLabel jLabel5;
-    private javax.swing.JLabel jLabel6;
-    private javax.swing.JLabel jLabel7;
-    private javax.swing.JLabel jLabel8;
-    private javax.swing.JScrollPane jScrollPane2;
-    private javax.swing.JTextField lab_ncheque;
-    private javax.swing.JTextArea lab_nconcepto;
-    private javax.swing.JTextField lab_ndoc;
-    private javax.swing.JTextField lab_nfecha;
-    private javax.swing.JTextField lab_nmonto;
-    private javax.swing.JTextField lab_nnombre;
     // End of variables declaration//GEN-END:variables
 }
