--- src/test/java/com/nutrehogar/sistemacontable/application/service/BackupServiceTest.java
@@ -1,22 +1,0 @@
-//package com.nutrehogar.sistemacontable.application.service;
-//
-//import com.nutrehogar.sistemacontable.persistence.config.HibernateUtil;
-//import org.junit.jupiter.api.BeforeEach;
-//import org.junit.jupiter.api.Test;
-//
-//import static org.junit.jupiter.api.Assertions.*;
-//
-//class BackupServiceTest {
-//    BackupService backupService;
-//    @BeforeEach
-//    void setUp() {
-//        backupService = BackupService.getInstance();
-//    }
-//
-//    @Test
-//    void test(){
-//        HibernateUtil.getSession();
-//    }
-//
-//
-//}
