--- src/test/java/com/nutrehogar/sistemacontable/application/service/ConfigLoaderTest.java
@@ -1,33 +1,0 @@
-//package com.nutrehogar.sistemacontable.application.service;
-//
-//import org.junit.jupiter.api.Test;
-//
-//
-//class ConfigLoaderTest {
-//
-//    @Test
-//    void test() {
-//        ConfigLoader configLoader = ConfigLoader.getInstance();
-//    }
-//
-//    @Test
-//    void getProperty() {
-//        for (ConfigLoader.Property property : ConfigLoader.Property.values()) {
-//            System.out.println(ConfigLoader.getProperty(property));
-//        }
-//
-//    }
-//
-//    @Test
-//    void getDbPath() {
-//    }
-//
-//    @Test
-//    void getBackupPath() {
-//    }
-//
-//    @Test
-//    void saveProperties() {
-//        ConfigLoader.saveProperties();
-//    }
-//}
